pyproject.toml
[tool.poetry]
name = 'K1st-Contrib'
version = '23.3.27'
description = 'K1st Contrib'
authors = ['Aitomatic, Inc. <engineering@aitomatic.com>']
license = 'Apache-2.0'
readme = 'README.md'
homepage = 'https://github.com/h1st-ai/h1st-contrib'
repository = 'https://github.com/h1st-ai/h1st-contrib'
packages = [
  {include = 'k1st_contrib'}
]


[tool.poetry.dependencies]
Python = '^3.8'

# H1st
H1st = {url = 'https://github.com/h1st-ai/h1st/archive/oracle-pmfp.zip'}

# Image Processing
Pillow = '>= 9.4.0'

# Hugging Face
Transformers = '>= 4.27.3'

# Django & related
Django = '>= 4.1.7'
Django-Configurations = '>= 2.4'
Django-CORS-Headers = '>= 3.14.0'
Django-Environ = '>= 0.10.0'
Django-Extensions = '>= 3.2.1'
DjangoRESTFramework = '>= 3.14.0'
Django-Silk = '>= 5.0.3'
DRF-Writable-Nested = '>= 0.7.0'

# DB-related
PsycoPG2-binary = '>= 2.9.5'


[build-system]
requires = [
  'Poetry-Core >= 1.5.2',
]
build-backend = 'poetry.core.masonry.api'
